utils/requirements.txt
streamlit>=1.32.0
streamlit-authenticator==0.3.2
pytesseract>=0.3.10
Pillow>=10.0.0
bcrypt>=4.0.0
PyYAML>=6.0
